租户后台导航栏充值提现未统计数量

See merge request game-plat-group/gameplat-admin-server!150

+++ src/main/java/com/gameplat/admin/service/RechargeOrderService.java
@@ -71,5 +71,5 @@ public interface RechargeOrderService extends IService<RechargeOrder> {
 
   List<ThreeRechReportVo> findThreeRechReport(GameRWDataReportDto dto);
 
-  Integer getUntreatedRechargeCount();
+  long getUntreatedRechargeCount();
 }
